Fixed warnings
git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@2571 fa19cda8-b258-4bce-aa1a-843e42da0b19

+++ src/sccp_config.c
@@ -1609,7 +1609,7 @@ sccp_device_t *sccp_config_applyDeviceConfiguration(sccp_device_t * d, struct as
 	uint16_t instance = 0;
 
 	char *splitter;
-	uint8_t digittimeout;
+	int digittimeout;
 
 	char config_value[256];
